wdm: Run IO work items in the main event loop thread
This commit changes wdm.dll to run IO work items in the main event
loop thread, instead of their dedicated thread. This is due to the
fact that IO work items can call IoCallDriver, and therefore may
need to run in a coroutine stack.

We also fix several bugs in the IRP forwarding logic. In particular
the Device object in an IRP (both server side and client side) now
refers to the current level device object in the device stack.