private/wdm: Use RIP-relative addressing in amd64/coroutine.S
Use RIP-relative relocation to address global variables in
private/wdm/src/amd64/coroutine.S. LLD 12 seems to incorrectly
strip the relocation section of the final PE file of wdm.dll so
accessing global variables using absolute addresses produces VM
faults. LLD 13 seems to have fixed this problem, but we want to
be compatible with both. RIP-relative relocation is the standard
way of addressing global variables in amd64 anyway so we should
be using it.