i386: Fix SEH bug
On i386, RtlUnwind calls RtlpCaptureContext to save the control
registers of its caller so that once stack unwinding is finished
it can jump back to its caller. This relies on the compiler NOT
optimizing the base pointer away for RtlUnwind. Unfortunately on
Clang/LLVM there isn't an easy way to control frame pointer omit
on a per-function basis. We solve this by calling compiler builtins,
such as __builtin_frame_address and __builtin_return_address.

Also add comments for i386 SEH. The code was not very well documented.