wdm/dma.c: Fix incorrect map register physical address
We didn't compute the map register base physical address correctly.
This commit fixes that. With this commit the floppy driver is now
fully functional and can read the entire floppy without error.

Also fix a bad coding style in wdm/dma.c. Width16Bits is defined
twice. Once in the public header and once in the wdm/dma.c code.