wdm: Rename the master header name wdm.h to ntddk.h
I feel that ntddk.h is a better name for the master header of the
device driver interface. "WDM" originally came from "Windows
driver model" which our driver interface is based on. I have now
taken to call it "Well-organized Driver Model" (thank you ChatGPT)
because Windows is a trademark of Microsoft. In the public interface
we shall now avoid the name "WDM" as that is somewhat confusing.
We are not Windows, after all, and the driver interface is not
exactly that of the WinNT5-era Windows driver model. "NTDDK" can
stand for Neptune OS Driver Development Kit.