Implement file-based caching
We switch to a design where caching is based on file blocks (just
like on Windows). The problem of aliasing is handled by the cache
manager by recording the file offset mapping between the FS file
objects and their underlying volume object. When dirty data are
being flushed, the cache manager willcarefully flush only the
mapped data region. This design requires that the file system does
not modify the sectors on the volume device that belong to the
FS-level file object (or at least not without going through the
Cc interfaces).

This involves some fairly extensive changes to Mm, especially when
it comes to image section objects. In particular, we refactor Mm to
use the cache manager interface to load, parse, and map the image
section objects. Additionally, due to the complexity of walking the
import table server-side, we have moved DLL loading client side,
and have made driver loading functions async, so they can wait for
the file system to load the driver images from disk.