wdm: pass short IO buffers in the IRP
For short input data, instead of mapping and unmapping the user
IO buffers, we copy them through service message area. This commit
implements the necessary server-side and client-side changes to
accomplish this. Note that since IO is asynchronous, output data
will need a special IO APC to copy to client-side which is yet to
be implemented. Also, if input data are in the service message
area, they MUST be copied regardless of the IO transfer type of
the target device, as the data in the service message area are
transient. DIRECT_IO cannot map the service message area so IO
buffers for device objects with DIRECT_IO must be at least 512
bytes.