ntos/mm: Mark the untyped as Requested as soon as giving it out
Previously we relied on whether the UNTYPED cap has any derived
caps to test whether this untyped has been given out to clients
of the Mm subsystem. However this is incorrect during the (however
brief) period of time when MmRequestUntyped has been called but
the client has not yet allocated the child node and inserted that
child node under the UNTYPED cap. This leads to a crash when the
system is low on memory because the allocation of the child node
can fail, and as a result of the allocation failure the ExPool
may indirectly call MmReleaseUntyped to release the UNTYPED that
the ExPool has requested for more pool pages. If this UNTYPED
happens to be the sibling of the previously handed out UNTYPED,
that UNTYPED would be incorrectly cleaned up as well due to the
recursive nature of MmReleaseUntyped. This commit fixes this bug
by marking the UNTYPED as Requested as soon as MmRequestUntyped
returns, so MmReleaseUntyped will not clean it up.