ntos: Implement NtWriteFile
Implement the general semantics of NtWriteFile, which includes
writing to the end of the file (thus changing the file size).
We move the per-view FILE_OFFSET_MAPPINGs into one single AVL tree
in the FCB, and try to minimize the number of IRPs to send when
loading data from the file systems. When the client driver changes
the file size as a response to a WRITE IRP, we update the server-
side information. WRITE is protected by a lock (on the server-side)
so no concurrent WRITE is allowed. Client side does not need to
worry about this (so no more ERESOURCEs needed) as the server makes
sure a WRITE is always sent after the previous WRITE is completed.