ntos: Implement symbolic link object
Add a new object type OBJECT_SYMBOLIC_LINK and implement the creation
and parse procedure for the symbolic link object. Add a new system
service NtCreateSymbolicLinkObject. Finally we add a new member to
the object header so the RemoveProc for the directory objects no
longer needs to do a search in order to find the parent link.