ntos/ob: Keep parsing even though remaining path is empty.
Previously we used the fact that the remaining path is empty to stop
parsing. This turns out to be inadequate when REPARSE is involved,
so keep parsing even though the remaining path is empty. We now
stop when the parse routine returns error (including the case where
the parse routine does not exist) or returns STATUS_NTOS_STOP_PARSING.