ntos: Allow parsing with an empty string
We need to distinguish the cases of opening the root directory of
a volume and opening the volume itself (which results in a file
object representing the entire volume linearly), so allow the
parse routines to parse an empty string. This requires some fairly
extensive changes to the way that the path separator is skipped.