ob: Implement NtClose
Implement NtClose by adding the close routine for object types.
When closing a file object, we send the IRP_MJ_CLEANUP to the
driver so it can do the proper cleanup. When deleting a file
object, we send the CloseFile notification so the driver will
delete its client-side handle and then invoke the IRP_MJ_CLOSE
routine, which will dereference the client object.