ntos/io: Implement NtDeleteFile
Implement NtDeleteFile by opening the file with FILE_DELETE_ON_CLOSE
flag. If the file is already opened, send the driver a new CREATE
IRP but make sure the FCB is shared with the master file object.
This is also done when the client is opening the same file wth a
different access mask.