wdm: Wait for master IRP to be sent before sending associated IRP
Previously we don't wait for the master IRP to be sent to the server
if the master IRP originates from the server. This is not correct
since the master IRP may carry the file size information, which
the server needs to see in order to correctly process the associated
IRPs.