ntos: clean up properly when driver crashes
When a driver crashes, make sure all IRPs queued on the driver are
completed with error status and all IRPs issued by the driver are
canceled (this is currently a stub and will be finished later).
Forcibly remove all its device objects and make sure the system is
still in a consistent state. The client can later request to restart
the crashed driver.