ntos/io: detach the IO packet if waiting failed
If for some reason the IO packet is queued to a driver (by
IopQueueIoPacket) but the subsequent wait failed (say due to OOM),
make sure we detach the IO packet from the driver IO packet queue.
Otherwise when the IO packet is freed the driver IO packet queue
will contain references to freed objects.