ntos/mm: find the correct cap node sibling
MiCapTreeGetNextSibling previously returns NULL if the given cap
node is the last sibling. This commit changes its behavior to
return the first sibling unless the given cap is the only child,
in which case we still return NULL. This is so that MmReleaseUntyped
can correctly release the other untyped even if we pass the second
sibling untyped to it.