wdm: fix bug in MmGetMdlPhysicalAddress
In C the + operator has higher precedence than & so it will be
evaluated first. The reason this bug did not manifest previously
is because we had a previous bug which returned the large page
size instead of the page size (see the commit 55dee0), so physical
addresses below 2MB (or 1MB on amd64) would still be correct.