ntos: Switch to a two-level CSpace for client threads
In order to implement the NT Local Procedure Call securely, the
original one-level CSpace shared between all threads within a
process does not suffice, since a thread can call on the endpoint
cap of another thread (within the same process) and impersonate
as that thread in terms of LPC communication with the server. We
change this design to a two-level CSpace where the outer CNode
contains thread-private caps, with its zeroth slot pointing to
a process-wide shared CNode. The client-side caps are guarded with
the thread ID so no conflict of handles can occur. The added benefit
of this design is that the maximum number of threads per process
is no longer constrained by the size of the process CNode.