pci: implement query resource requirements
Implement the query resource requirements PNP request for the PCI
driver, by building the resource requirements list from both the
current resources that have already been assigned by (usually) the
boot firmware, as well as the port or memory range limits obtained
from probing. The former is set as the preferred resources and
the latter as the alternatives. In order to keep track of the BAR
index from which the resource is derived, we pass it as device
private data which the PNP manager will pass along verbatim.