storport: implement timer and DPC
Implement the storport timer and STOR_DPC. We elect to use an IO
work item for both in order to simplify locking. The timer will
always be a low-priority timer and the STOR_DPC is in fact just an
IO work item. This way no locking is required for the objects that
are accessed by the timer callbacks and DPC routines.