hal: enable IO APIC and use HPET for system timer
Use IO APIC instead of the legacy PIC so we can have more interrupt
pins. We don't support interrput sharing due to the context switching
cost of doing so, so it is important we have sufficient interrupt
pins. We also switch to using HPET for the system timer, rather than
the legacy PIT.