fatfs: reject the mount if data area is not cluster-aligned
Due to the way our cache manager works, all data access from the
file system driver must be cluster-aligned. On some small FAT12/16
volumes the data area may start at a non-cluster-aligned sector.
Since the likelihood of this is small, we reject the mount. Users
are advised to reformat the volume with a cluster-aligned data area
or set the cluster size to the sector size.