hal: use PIT for i386
QEMU seems to mis-emulate HPET on i386 in some cases so we revert
to using the PIT on i386 by default. On amd64 we still use HPET for
the system timer. The default can be changed by changing a global
variable, which we eventually want to make into a registry config
once we implemented the on-disk format of the registry.