wdm: fix incorrect SL_COMPLETION_STOPPED checks
The SL_COMPLETION_STOPPED control flag should be set in the same
IO stack location as the completion routine. If a IO completion
routine has requested the IO processing to be stopped, the stack
location should point to the one immediately above the IO stack
location in which the completion routine is registered. Before the
finalizer is run, it should therefore check the next stack location
for the control flags, rather than the current one.