wdm: save file object after IRP completion and deref later
We need to clear the device object and file object members of an
IO stack before calling the IO completion routine. When we are in
the last IO stack location, we need to save the file object and
deref it later so we can send the server the file size information.