ntos: insert file object into subobject directory using full path
Previously we have not been inserting the file object into the
subobject directory correctly which results in parsing errors when
the client accesses paths with multiple levels of subdirectories.
This commit fixes this by creating the sub-directories of the
subobject directory and inserting the file object using the final
component of the object path.