io: ensure consistent refcounting and lifetime of device objects
Make sure the NT Executive and the client driver processes see a
consistent set of device objects and their global handles. When
we grant a device object's global handle to a foreign driver, the
device object's refcount is increased by one, and a CloseDevice
message is preallocated. When the owning driver of the device object
deletes the device object, the CloseDevice message is sent to all
the foreign driver which has its global handle granted, and the
client will cleanup its local bookkeeping of the device object and
call WdmDeleteDevice to notify the server that it no longer expects
the global handle to be valid, at which point the server will deref
the (server-side) device object. This way the lifetime of device
objects is tracked consistently across the server and client drivers,
and the server can freely reuse its global handle after the refcount
of the device object has reached zero.