fatfs: do not deny creating a directory file
Do not deny creating a directory file even if FILE_DIRECTORY_FILE
is not specified. The caller is not required to specify either
FILE_DIRECTORY_FILE or FILE_NON_DIRECTORY_FILE. When neither is
specified, it means that the caller is ok with opening both non-
directory files and directory files.