fatfs: do not increase refcount of FCB when init caching
For some reason the ReactOS driver referenced the FCB when it
initializes the caching for the FCB of a directory. This creates
a refcounting leak as this reference is never decreased. Later
when responding to a IRP_MJ_CLEANUP, we need to deref the stream
file object if the file object being deleted is a directory.