ntos/io: allow function driver to be loaded in its bus driver process
Add a DWORD value under the enum key of a device instance named
LoadIntoBusDriver. If set to TRUE, the function driver of this
device instance will be loaded into the same driver process as
its bus driver. This is useful for function drivers that have
heavy communications with its bus driver, so they can share the
same address space to reduce the number of context switches. An
example of this is the storage class driver disk.sys and partition
manager partmgr.sys, which are the function drivers of the storage
disk devices and partition devices. Prior to this change, the port
driver generates a SCSI\<Bus> hardware ID for the port devices it
enumerates, and disk.sys and partmgr.sys are loaded as filter
drivers on top of the port driver. This breaks the Windows semantics
of PnP driver stack as disk.sys is really the function driver of the
port device, rather than a filter driver. With this change, we no
longer need the SCSI\<BUS> hack and can just load the function driver
disk.sys into the port device function driver (say storahci.sys),
enabling us to implement the same semantics as Windows.