wdm: disable IO port range if requested length does not match
If the requested IO port range does not match what we already have,
disable the existing overlapping ranges and request new range. For
instance, if acpi.sys already has enabled port 0x400, 0x401 (each
of which is an 8-bit port) and wants to access port 0x400 but as a
16-bit port, we disable port 0x400 and 0x401 and re-enable 16-bit
port 0x400.